chore: added unit test for resetPassword use case
