+++ __mocks__/src/containers/AppContainer.js
@@ -35,4 +35,5 @@ export default {
   getRetrieveEmailAndHashedPasswordGateway: jest.fn(() => () =>
     Promise.resolve({})
   ),
+  getResetPasswordGateway: jest.fn(() => () => Promise.resolve({})),
 };
